% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{vs_example_file}
\alias{vs_example_file}
\title{Example DataVolley files provided as part of the volleysim package}
\usage{
vs_example_file(choice = 1)
}
\arguments{
\item{choice}{numeric: which data file to return?
\itemize{
\item{1 - the 2020 Austrian Women's Volley Cup played between Hartberg and UVC Graz}
}}
}
\value{
path to the file
}
\description{
Example DataVolley files provided as part of the volleysim package
}
\examples{
\dontrun{
  myfile <- vs_example_file()
  x <- datavolley::dv_read(myfile)
  summary(x)
}

}
\references{
The example data files came from \url{https://www.volleynet.at/dvdownload/information/f-Damen/}
}
\seealso{
\code{\link[datavolley]{dv_read}}
}
