% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoretical.R
\name{vs_theoretical_sideout_rates}
\alias{vs_theoretical_sideout_rates}
\title{Estimate theoretical sideout rates given 'phase' parameters}
\usage{
vs_theoretical_sideout_rates(rates, process_model = "phase")
}
\arguments{
\item{rates}{list: rates as returned by \code{\link[=vs_estimate_rates]{vs_estimate_rates()}}}

\item{process_model}{string: currently only "phase". See \code{\link[=vs_estimate_rates]{vs_estimate_rates()}}}
}
\value{
The theoretical sideout rates of the two teams
}
\description{
The \code{\link[=vs_estimate_rates]{vs_estimate_rates()}} function returns a team's performance rates across a range of aspects of play, including serve ace rate, serve error rate, and so on. Using \code{\link[=vs_theoretical_sideout_rates]{vs_theoretical_sideout_rates()}} We can estimate the theoretical sideout rate that we would expect to see, given those parameters. This can be compared to the actual sideout rate achieved by the team.
}
\examples{
\dontrun{
  library(datavolley)
  x <- dv_read(dv_example_file())
  rates <- list(vs_estimate_rates(x, target_team = home_team(x)),
                vs_estimate_rates(x, target_team = visiting_team(x)))

  ## the theoretical sideout rates
  vs_theoretical_sideout_rates(rates)

  ## compare to their actual sideout rates
  c(rates[[1]]$sideout, rates[[2]]$sideout)
}
}
\seealso{
\code{\link[=vs_estimate_rates]{vs_estimate_rates()}}
}
