% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{vs_estimate_rates}
\alias{vs_estimate_rates}
\title{Estimate parameters required for simulation}
\usage{
vs_estimate_rates(x, target_team, by = "none")
}
\arguments{
\item{x}{datavolleyplays: the plays component of a datavolley object as returned by \code{\link[datavolley]{dv_read}}}

\item{target_team}{string: the team name to calculate rates for. If missing or NULL, rates will be calculated across the entire data.frame \code{x}. If \code{target_team} is "each", rates will be calculated for each team separately}

\item{by}{string: grouping to calculate rates by. Either "none" (calculate whole-data set rates), "match" (by match), or "set" (by match and set)}
}
\value{
A tibble, currently with the columns sideout, serve_ace, serve_error, rec_set_error, rec_att_error, rec_att_kill, trans_set_error, trans_att_error, trans_att_kill, rec_block, and trans_block, plus (if \code{by} is "match") match_id and (if \code{by} is "set") set_number and (if \code{target_team} is "each") "team"
}
\description{
Estimate parameters required for simulation
}
\examples{
\dontrun{
  library(datavolley)
  x <- dv_read(dv_example_file())
  rates <- list(vs_estimate_rates(x, target_team = home_team(x)),
                vs_estimate_rates(x, target_team = visiting_team(x)))
  vs_simulate_set(rates) ## simulate a single set
  vs_simulate_match(rates, n = 100, simple = TRUE) ## simulate a match 100 times
  ## so we expect the home team to win, with 3-0 being the most likely scoreline

  ## compare to the actual match result
  summary(x)
}
}
\seealso{
\code{\link{vs_simulate_set}}
}
