% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{vs_simulate_match}
\alias{vs_simulate_match}
\title{Simulate a volleyball match}
\usage{
vs_simulate_match(
  rates,
  process_model = "phase",
  serving = NA,
  n = 2000,
  simple = FALSE
)
}
\arguments{
\item{rates}{list: A two-element list, each element of which is a set of rates as returned by \code{vs_estimate_rates}}

\item{process_model}{string: either "sideout" or "phase". Details TBD}

\item{serving}{logical: if \code{TRUE}, team 1 will serve first. If \code{NA}, the team serving first will be chosen at random}

\item{n}{integer: the number of simulations to run}

\item{simple}{logical: if \code{TRUE}, just return the probability of team winning and the probabilities of each possible set score. If \code{FALSE}, return extra details in a named list}
}
\description{
Currently hard-coded to indoor, best-of-5-set scoring.
}
\examples{
\dontrun{
  library(datavolley)
  x <- dv_read(dv_example_file())
  rates <- list(vs_estimate_rates(x, target_team = home_team(x)),
                vs_estimate_rates(x, target_team = visiting_team(x)))
  vs_simulate_set(rates) ## simulate a single set
  vs_simulate_match(rates, n = 100, simple = TRUE) ## simulate a match 100 times
  ## so we expect the home team to win, with 3-0 being the most likely scoreline

  ## compare to the actual match result
  summary(x)
}

}
\seealso{
\code{\link{vs_estimate_rates}} \code{\link{vs_simulate_match}}
}
