% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{vs_simulate_set}
\alias{vs_simulate_set}
\title{Simulate a set of volleyball}
\usage{
vs_simulate_set(rates, serving = NA, go_to = 25, simple = FALSE, id = NULL)
}
\arguments{
\item{rates}{list: A two-element list, each element of which is a set of rates as returned by \code{vs_estimate_rates}}

\item{serving}{logical: if \code{TRUE}, team 1 will serve first. If \code{NA}, the team serving first will be chosen at random}

\item{go_to}{integer: the minimum score that must be reached to end the set (typically 25 for indoor volleyball in sets 1 to 4, 15 in set 5, or 21 in beach volleyball)}

\item{simple}{logical: if \code{TRUE}, just return the team (1 or 2) that won. If \code{FALSE}, return extra details in a data.frame}

\item{id}{: an optional value that (if non-\code{NULL}) will be returned in the \code{id} column of the returned data frame, if \code{simple} is \code{FALSE}}
}
\value{
Integer (1 or 2) or a data frame, depending on the value of \code{simple}
}
\description{
Simulate a set of volleyball
}
\examples{
\dontrun{
  library(datavolley)
  x <- dv_read(dv_example_file())
  rates <- list(vs_estimate_rates(x, target_team = home_team(x)),
                vs_estimate_rates(x, target_team = visiting_team(x)))
  vs_simulate_set(rates) ## simulate a single set
  vs_simulate_match(rates, n = 100, simple = TRUE) ## simulate a match 100 times
  ## so we expect the home team to win, with 3-0 being the most likely scoreline

  ## compare to the actual match result
  summary(x)
}

}
\seealso{
\code{\link{vs_estimate_rates}} \code{\link{vs_simulate_match}}
}
